refactor: Update FilesTreeView constructor to include SvgWidget parameter
The FilesTreeView constructor has been updated to include a SvgWidget parameter. This allows for the integration of the SvgWidget class into the FilesTreeView, enabling the loading of SVG files into the SvgWidget when a file is double-clicked in the tree view. This change improves the functionality and user experience of the UI design.