refactor: Update SvgHandler class to use loadSvg instead of changeSvg
This commit refactors the SvgHandler class in the UI design. The changeSvg function has been renamed to loadSvg to better reflect its purpose of loading an SVG file. The function now also includes additional logic to parse the loaded SVG file using the SystemcLinker class. These changes improve the clarity and consistency of the code.