refactor: Update SvgParser.h to include DebugWindow header file
This commit adds the necessary include statement for the DebugWindow header file in SvgParser.h. This change allows the SvgParser class to use the DebugWindow class for debugging purposes. By including the DebugWindow header file, the SvgParser class can access the necessary functions and variables from the DebugWindow class, improving the debugging capabilities of the project.