refactor: Add TEMP_SVG_ER_PATH constant for error SVG file
This commit adds the TEMP_SVG_ER_PATH constant to the params.h file, which specifies the path to the error SVG file used for debugging. The constant is used in the mainwindow.cpp file to load the error SVG when the corresponding button is clicked. This change improves the debugging capabilities of the application.