refactor: Connect QTreeWidget itemClicked signal to ElementsDisplay slot
This commit modifies the `ElementsDisplay` class in `elementsdisplay.cpp` and `elementsdisplay.h` files. It adds a new slot `handleItemClicked` and connects it to the `itemClicked` signal of the `QTreeWidget` in the constructor. This change allows for handling the click event on tree items and emitting the `elementClicked` signal with the clicked item's text. This will enable to high elements in the svgwidget display