refactor: Add buttons to SecondWindow for true/false input
This commit adds buttons to the SecondWindow class in secondwindow.cpp and secondwindow.h files. The trueButton and falseButton are created and added to the layout. Clicking the trueButton calls the handleTrueButton function, which adds a true point to the MutiTypesChrono widget. Clicking the falseButton calls the handleFalseButton function, which adds a false point. Debug only