feat: Add popup display for data points on graph
This commit adds the functionality to display a popup with the coordinates of a data point on the graph when the mouse cursor is near the point. The popup is styled with a white background, black border, and padding. It is hidden by default and only shown when the cursor is close to a data point. This feature enhances the user experience by providing more information about the data points on the graph.