refactor: Improve thread safety in SvgHandler class
This commit introduces a static mutex and a static pointer to the last created instance in the `SvgHandler` class. The mutex is used to protect access to a static list of components, and the pointer is used to automatically retrieve the last parsed components list. This improves thread safety and ensures that the correct components list is returned even when multiple instances of `SvgHandler` are created.