refactor: Update MultiTypesChrono class to support multiple data types
This commit updates the MultiTypesChrono class to support multiple data types for data points on the graph. It introduces a template parameter T and adds a new method addPoint(int point) to add integer data points to the graph. This refactor improves code flexibility and allows for visualizing different types of data on the graph.