refactor: Update MultiTypesChrono class to support multiple data types
This commit updates the MultiTypesChrono class to support multiple data types for data points on the graph. It introduces a template parameter T and adds new methods addPoint(int point), addPoint(float point), and addPoint(double point) to add different types of data points to the graph. This refactor improves code flexibility and allows for visualizing different types of data on the graph.